+++ plugin/src/sbt-test/ScalaTsiPlugin/negative-tests/expected_model_scala2.ts
@@ -1,0 +1,13 @@
+// DO NOT EDIT: generated file by scala-tsi
+
+export interface IClassWithUndefinedMember {
+  notDefined: "Could not find TSType[models.NotDefined] in scope and could not generate it"
+}
+
+export interface IRecursiveA {
+  b: IRecursiveB
+}
+
+export interface IRecursiveB {
+  a: IRecursiveA
+}