+++ plugin/src/sbt-test/ScalaTsiPlugin/negative-tests/src/main/scala/models/ClassWithUndefinedMember.scala
@@ -1,0 +1,7 @@
+package models
+
+case class ClassWithUndefinedMember(
+    notDefined: NotDefined
+)
+
+class NotDefined {}