+++ src/main/scala/com/scalatsi/DefaultTSTypes.scala
@@ -3,8 +3,7 @@ package com.scalatsi
 /** The default Scala-to-Typescript mappings provided by this library */
 // Cannot be put in the scala-agnostic code because 2.12 does not support literal types
 trait DefaultTSTypes
-    extends PrimitiveTSTypes
-    with ScalaTSTypes
+    extends ScalaTSTypes
     with CollectionTSTypes
     with FunctionTSTypes
     with TupleTSTypes